[refactor] #10 폴더 이름 변경 ( product -> funding )
