+++ src/main/java/com/example/shipgofunding/funding/repository/FundingJpaRepository.java
@@ -1,6 +1,6 @@
-package com.example.shipgofunding.product.repository;
+package com.example.shipgofunding.funding.repository;
 
-import com.example.shipgofunding.product.domain.Funding;
+import com.example.shipgofunding.funding.domain.Funding;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FundingJpaRepository extends JpaRepository<Funding, Integer> {
